Add `to_gpu` and `to_cpu` support for both tensor and DataFrame (#630)
(cherry picked from commit b28db59af5bf4dbf6b4066b6c23a739648bc8a1b)
