+++ mars/dataframe/base/tests/test_base.py
@@ -1,0 +1,83 @@
+# Copyright 1999-2018 Alibaba Group Holding Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+import pandas as pd
+
+from mars.tests.core import TestBase
+from mars.dataframe.datasource.dataframe import from_pandas as from_pandas_df
+from mars.dataframe.datasource.series import from_pandas as from_pandas_series
+from mars.dataframe.base import to_gpu, to_cpu
+
+
+class Test(TestBase):
+    def testToGPU(self):
+        # test dataframe
+        data = pd.DataFrame(np.random.rand(10, 10), index=np.random.randint(-100, 100, size=(10,)),
+                            columns=[np.random.bytes(10) for _ in range(10)])
+        df = from_pandas_df(data)
+        cdf = to_gpu(df)
+
+        self.assertEqual(df.index_value, cdf.index_value)
+        self.assertEqual(df.columns, cdf.columns)
+        self.assertTrue(cdf.op.gpu)
+        pd.testing.assert_series_equal(df.dtypes, cdf.dtypes)
+
+        cdf.tiles()
+
+        self.assertEqual(df.nsplits, cdf.nsplits)
+        self.assertEqual(df.chunks[0].index_value, cdf.chunks[0].index_value)
+        self.assertEqual(df.chunks[0].columns, cdf.chunks[0].columns)
+        self.assertTrue(cdf.chunks[0].op.gpu)
+        pd.testing.assert_series_equal(df.chunks[0].dtypes, cdf.chunks[0].dtypes)
+
+        self.assertIs(cdf, to_gpu(cdf))
+
+        # test series
+        sdata = data.iloc[:, 0]
+        series = from_pandas_series(sdata)
+        cseries = to_gpu(series)
+
+        self.assertEqual(series.index_value, cseries.index_value)
+        self.assertTrue(cseries.op.gpu)
+
+        cseries.tiles()
+
+        self.assertEqual(series.nsplits, cseries.nsplits)
+        self.assertEqual(series.chunks[0].index_value, cseries.chunks[0].index_value)
+        self.assertTrue(cseries.chunks[0].op.gpu)
+
+        self.assertIs(cseries, to_gpu(cseries))
+
+    def testToCPU(self):
+        data = pd.DataFrame(np.random.rand(10, 10), index=np.random.randint(-100, 100, size=(10,)),
+                            columns=[np.random.bytes(10) for _ in range(10)])
+        df = from_pandas_df(data)
+        cdf = to_gpu(df)
+        df2 = to_cpu(cdf)
+
+        self.assertEqual(df.index_value, df2.index_value)
+        self.assertEqual(df.columns, df2.columns)
+        self.assertFalse(df2.op.gpu)
+        pd.testing.assert_series_equal(df.dtypes, df2.dtypes)
+
+        df2.tiles()
+
+        self.assertEqual(df.nsplits, df2.nsplits)
+        self.assertEqual(df.chunks[0].index_value, df2.chunks[0].index_value)
+        self.assertEqual(df.chunks[0].columns, df2.chunks[0].columns)
+        self.assertFalse(df2.chunks[0].op.gpu)
+        pd.testing.assert_series_equal(df.chunks[0].dtypes, df2.chunks[0].dtypes)
+
+        self.assertIs(df2, to_cpu(df2))
