+++ mars/opcodes.py
@@ -25,6 +25,8 @@ DATAFRAME_DATA_SOURCE = 17
 DATAFRAME_FROM_TENSOR = 21
 DATAFRAME_FROM_RECORDS = 24
 SERIES_DATA_SOURCE = 23
+TO_GPU = 30
+TO_CPU = 31
 RAND_RAND = 41
 RAND_RANDN = 42
 RAND_RANDINT = 43
