+++ mars/serialize/protos/operand.proto
@@ -39,6 +39,10 @@ message OperandDef {
         // series
         SERIES_DATA_SOURCE = 23;
 
+        //GPU
+        TO_GPU = 30;
+        TO_CPU = 31;
+
         // random
         RAND_RAND = 41;
         RAND_RANDN = 42;
