+++ mars/tensor/__init__.py
@@ -50,6 +50,7 @@ from .reshape import reshape
 from .merge import concatenate, stack, hstack, vstack, dstack, column_stack
 from .rechunk import rechunk
 from .lib.index_tricks import mgrid, ogrid, ndindex
+from .core import mutable_tensor
 
 from . import random
 from . import fft
